Face_Anti_Spoofing/requirements.txt
absl-py                          0.15.0
arviz                            0.11.2
astor                            0.8.1
astunparse                       1.6.3
backports.zoneinfo               0.2.1
baidu-aip                        4.16.13
bidict                           0.22.1
blessings                        1.7
Bottleneck                       1.3.5
cached-property                  1.5.2
cachetools                       5.3.2
certifi                          2023.11.17
cftime                           1.5.1.1
chardet                          5.2.0
charset-normalizer               3.3.2
click                            8.1.7
cmake                            3.29.3
cmdstanpy                        0.9.5
colorama                         0.4.6
convertdate                      2.3.2
cycler                           0.11.0
Cython                           0.29.33
dlib                             19.19.0
ephem                            4.1.2
et-xmlfile                       1.1.0
fbprophet                        0.7.1
Flask                            2.2.5
Flask-Cors                       4.0.1
Flask-SocketIO                   5.3.6
flatbuffers                      1.12
flit_core                        3.6.0
freetype-py                      2.4.0
gast                             0.2.2
google-auth                      2.25.1
google-auth-oauthlib             0.4.6
google-pasta                     0.2.0
gpustat                          0.4.1
grpcio                           1.37.1
h11                              0.14.0
h5py                             2.10.0
hijri-converter                  2.2.2
holidays                         0.18
idna                             3.6
importlib-metadata               6.7.0
imutils                          0.5.4
itsdangerous                     2.1.2
jieba                            0.42.1
Jinja2                           3.1.2
Keras                            2.2.4
Keras-Applications               1.0.8
keras-bert                       0.80.0
keras-embed-sim                  0.8.0
keras-layer-normalization        0.14.0
keras-multi-head                 0.27.0
keras-pos-embd                   0.11.0
keras-position-wise-feed-forward 0.6.0
Keras-Preprocessing              1.1.2
keras-self-attention             0.46.0
keras-transformer                0.30.0
kiwisolver                       1.4.4
korean-lunar-calendar            0.2.1
LunarCalendar                    0.0.9
Markdown                         3.4.4
MarkupSafe                       2.1.3
matplotlib                       3.1.3
mkl-fft                          1.3.1
mkl-random                       1.2.2
mkl-service                      2.4.0
netCDF4                          1.5.3
numexpr                          2.8.4
numpy                            1.20.2
nvidia-ml-py3                    7.352.0
oauthlib                         3.2.2
opencv-python                    4.9.0.80
openpyxl                         3.0.10
opt-einsum                       3.3.0
packaging                        22.0
pandas                           1.3.5
Pillow                           9.5.0
pip                              22.3.1
plotly                           5.9.0
protobuf                         3.20.3
psutil                           5.8.0
pyasn1                           0.5.1
pyasn1-modules                   0.3.0
PyMeeus                          0.5.11
pyparsing                        3.0.9
pyreadline                       2.1
pystan                           2.19.1.1
python-dateutil                  2.8.2
python-engineio                  4.9.0
python-socketio                  5.11.0
python-version                   0.0.2
pytz                             2022.7
PyYAML                           5.4.1
requests                         2.31.0
requests-oauthlib                1.3.1
rsa                              4.9
scipy                            1.6.3
setuptools                       68.0.0
setuptools-git                   1.2
simple-websocket                 1.0.0
six                              1.15.0
tenacity                         8.0.1
tensorboard                      2.1.1
tensorboard-data-server          0.6.1
tensorboard-plugin-wit           1.8.1
tensorflow                       1.15.0
tensorflow-estimator             1.15.1
tensorflow-gpu                   2.4.0
tensorflow-gpu-estimator         2.1.0
termcolor                        1.1.0
tornado                          5.1.1
tqdm                             4.60.0
typing-extensions                3.10.0.0
urllib3                          1.26.16
Werkzeug                         2.2.3
wheel                            0.42.0
wincertstore                     0.2
wrapt                            1.12.1
wsproto                          1.2.0
xarray                           0.20.1
zipp                             3.15.0